fix: change table rotation from float to int, constrain to [0, 359]
Rotation is always a whole number of degrees (UI rotates in 15° steps).
- models.py: Mapped[float]/Float → Mapped[int]/Integer
- schemas.py: float → int, le=360 → le=359 in TableCreate, TableUpdate, TableOut
- migration 006: ALTER COLUMN rotation Float → Integer for existing DBs
- reservation.ts: update JSDoc comment
- tables.py: minor comment tweak

+++ backend/alembic/versions/001_initial.py
@@ -53,7 +53,7 @@ def upgrade() -> None:
         sa.Column("shape", sa.String(20), nullable=False, server_default="rectangle"),
         sa.Column("width_m", sa.Float, nullable=False, server_default="1.8"),
         sa.Column("height_m", sa.Float, nullable=False, server_default="0.7"),
-        sa.Column("rotation", sa.Float, nullable=False, server_default="0"),
+        sa.Column("rotation", sa.Integer, nullable=False, server_default="0"),
         sa.Column(
             "created_at",
             sa.DateTime(timezone=True),
